working on Wedding Testing

+++ weddingTherapy/src/main/java/com/revature/wedding_therapy/services/WeddingService.java
@@ -17,11 +17,12 @@ public class WeddingService {
 	
 	
 	public boolean createNewWedding(Weddings newWedding) throws WeddingNotNamedException {
-		if(newWedding.getWedding_name() == null || newWedding.getWedding_name().equals("")) {
+		if(!validWedding(newWedding)){
 			throw new WeddingNotNamedException("Please Give This Wedding A Name!");
 		}
 		return weddingDAO.createNewWedding(newWedding);
 	}
+	
 	
 	public List<Weddings> findAllWeddings(){
 		return weddingDAO.findAllWeddings();
@@ -29,6 +30,7 @@ public class WeddingService {
 	
 	
 	public Weddings getWedding(int wedding_id) {
+		if(!validId(wedding_id)) {return null;}
 		return weddingDAO.findWeddingByID(wedding_id);
 	}
 	
@@ -39,8 +41,25 @@ public class WeddingService {
 	
 	
 	public boolean deleteWedding(int wedding_id) {
-		//System.out.println("\nWeddingService:deleteWedding\n");
+		if(!validId(wedding_id)) {return false;}
 		return weddingDAO.deleteWedding(wedding_id);
 	}
+	
+	
+	private boolean validWedding(Weddings newWedding) {
+		if(newWedding.getWedding_name() == null || newWedding.getWedding_name().equals("") || 
+				newWedding.getWedding_date() == null || newWedding.getWedding_date().equals("") || 
+		newWedding.getWedding_budget() <= 0 ) {
+			return false;
+		}
+		return true;
+	}
 
+	
+	private boolean validId(int id) {
+		if(99999 >= id && id >= 1) {
+			return true;
+		}
+		return false;
+	}
 }
